+++ examples/data-sources/openaiagents_environment_template/data-source.tf
@@ -1,3 +1,7 @@
+resource "openaiagents_environment_template" "skills" {
+  name = "report-env"
+}
+
 data "openaiagents_environment_template" "selected" {
   id = openaiagents_environment_template.skills.id
 }